requirements.txt
Django==4.2.8
djangorestframework>=3.15.0,<3.15.2
psycopg2==2.9.9
drf-spectacular>=0.27.0,<0.27.2
django-ulid==0.0.4
ulid-py==1.1.0